tcp connect

+++ src/dacn_letrananhnhung/IDonHang.java
@@ -2,5 +2,5 @@
 package dacn_letrananhnhung;
 
 public interface IDonHang {
-    double thanhTien();
+    float thanhTien();
 }
